refactor TraceHandler.kt separate classes in different files add comments for important classes
